Ajout de la fonctionnalité de création de tickets avec mise à jour de l'interface utilisateur et intégration des clients dans le composant Test
